+++ src/main/java/com/konkuk/strhat/domain/self_diagnosis/entity/SelfDiagnosis.java
@@ -1,7 +1,7 @@
-package com.konkuk.strhat.self_diagnosis.entity;
+package com.konkuk.strhat.domain.self_diagnosis.entity;
 
-import com.konkuk.strhat.self_diagnosis.enums.SelfDiagnosisType;
-import com.konkuk.strhat.user.entity.User;
+import com.konkuk.strhat.domain.self_diagnosis.enums.SelfDiagnosisType;
+import com.konkuk.strhat.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
 import java.time.LocalDate;
