+++ src/main/java/com/konkuk/strhat/domain/user/enums/Gender.java
@@ -1,13 +1,13 @@
-package com.konkuk.strhat.chat.enums;
+package com.konkuk.strhat.domain.user.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum Sender {
-    USER("유저"),
-    CHAT_BOT("챗봇");
+public enum Gender {
+    MALE("남자"),
+    FEMALE("여자");
 
     private final String description;
 }
